Проходят снэпшот тесты для первых компонентов
