+++ src/components/review-form/review-form.jsx
@@ -21,12 +21,14 @@ const ReviewForm = (props) => {
   const handleSubmit = (evt) => {
     evt.preventDefault();
     setDisabledInput(true);
-    return new Promise(() => {
-      postReview(id, {text, rating});
+    setDisabledSubmit(true);
+    return new Promise((resolve, reject) => {
+      postReview(id, {text, rating}, resolve, reject);
     }).then(() => {
       setText(``);
       setRating(0);
       setDisabledInput(false);
+      setDisabledSubmit(false);
     }).catch(() => {
       setshowErrorMessage(true);
     });
