+++ src/index.js
@@ -1,30 +1,12 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import {createStore, applyMiddleware} from "redux";
 import {Provider} from "react-redux";
-import thunk from "redux-thunk";
-import {createAPI} from "./services/api";
-import {composeWithDevTools} from "redux-devtools-extension";
 import App from "./components/app/app";
 import {cities} from "./utils/const";
-import rootReducer from "./store/reducers/root-reducer";
-import {ActionCreator} from "./store/action";
-import {AuthorizationStatus} from "./utils/const";
 import {fetchOffersList} from "./store/api-actions";
-import {redirect} from "./middlewares/redirect";
 import ErrorPage from "./components/error-page/error-page";
+import store from "./store/store";
 
-const api = createAPI(
-    () => store.dispatch(ActionCreator.requireAuthorization(AuthorizationStatus.NO_AUTH))
-);
-
-export const store = createStore(
-    rootReducer,
-    composeWithDevTools(
-        applyMiddleware(thunk.withExtraArgument(api)),
-        applyMiddleware(redirect)
-    )
-);
 store.dispatch(fetchOffersList()).then(
 
     ReactDOM.render(
